1) views/functions.xsl исправлена функция форматирования сумм для правильной выгрузки в Excel и функция форматирования дат 2) Внебрачные контракты 	Пользоватекль: поле, разрешающее использование внебрачных контрактов  allow_ext_contracts 	user_view.sql 	user_dialog.sql 	UserDialog_View.js 	User_Controller.php новый глобальный парамтер allow_ext_contracts 	ViewBase_php.xsl allow_ext_contracts 	таблица applications поле ext_contract bool default FALSE 	applications_list.sql ТОЛЬКО ext_contract=FALSE 	Модель ApplicationExtList, полная копия ApplicationList только показывающая ext_contract=TRUE 	Application_Controller new methods: get_ext_list,complete_ext 	Новый пункт меню (views) заявление (внеконтрактные) 	ApplicationExtList_View.js MainMenuConstructor_Controller->update исправлен косяк при вызове функции без параметра 	ApplicationDialog_View.js 	ApplicationDialog_html.xsl 	application_processes_process.sql 	doc_flow_out_client_process.sql 	doc_flow_in_process.sql 	Таблица doc_flow_in поле ext_contract заполняется из doc_flow_in_process.sql 	Новая константа ext_contract_doc_pref 	Новый параметр в функции doc_flow_in_next_num, вызывается из doc_flow_in_process. Теперь также передается контракт/внеконтракт 	doc_flow_in_list.sql Условие только контракт ext_contract=FALSE 	doc_flow_in_ext_list.sql 	views + внеконтракт 	applications_ref.sql 	Таблица doc_flow_out новое поле ext_contract 	doc_flow_out_dialog.sql 	doc_flow_examinations_dialog.sql 	DocFlowExamination_View.js 	DocFlowOutDialog_View.js, тема 	Application_Controller.php,Contract_Controller.php функция get_constr_name() теперь возвращант и признак внеконтракта 	tmpl/DocFlowOutDialog.html 	DocFlowOut_Controller->get_next_num() теперь принимает и признак внеконтракта для формирования номера 	Новый параметр в функции doc_flow_out_next_num, вызывается из doc_flow_in_process. Теперь также передается контракт/внеконтракт 	doc_flow_in_dialog.sql 	BtnNextContractNum.js 	DocFlowOut_Controller->get_next_contract_number() теперь принимает и признак внеконтракта для формирования номера 	contracts_next_number.sql 	doc_flow_out_list.sql Только контракт! 	doc_flow_out_ext_list.sql 	views Исходящие документы (внеконтракт) 	Новое представление для журнала исх.жлкументов внеконтракт DocFlowOutExtList_View.js 	contracts_list.sql Только контракты! 	views Контракты (внеконтракты) 	forms/ContractExtList_Form.js 	views/ContractExtList_View.js 	doc_flow_examinations_process.sql 	contracts_dialog.sql 	ContractDialog_View.js 	ContractDialog.html 	contracts_expertise_result_number.sql 	document_templates_list ДОБИТЬ функцию contracts_ext_to_contract.sql сталось: 	+Исходящие документы (список,поле, создание из списка с признаком или на форме) 	+Создание контракта (свой список, префикс) 	+Копирование всех документов из внеконтракт ->> контракт 	+Убрать из меню у кого не надо !!!vendor/cakephp/filesys/Folder line 580,curly braces
+Статусы заявления на доступ админу. Все формы статусы.
	Сделано ранне, проверил
+Копирование старой достоверности, не переходлит в новую услугу
	Application_Controller.php
+При создании заявления по достоверности, записали, открыли, появляется обязательное поле Повторное использование/неиспользование... и требует
его заполнения!!!
	ApplicationDialog_View.js
+Разные сроки блокировки отправки исх.писем по услугам.
поле ban_client_responses_day_cnt в services, ServiceList_View.js, DocFlowOutClient_Controller->check_resp_doc()

+Исходящее наше письмо, добавить возможность у папок, которые требуют подписания клиента, убирать галочку, или наоборот,
чтобы при вставке в таблицу application_document_files сразу ставилась галочка file_signed_by_client
Поле doc_flow_attachments.require_client_sig
doc_flow_out_dialog
doc_flow_in_client_dialog
doc_flow_out_client_files_for_signing
FileUploaderDocFlowOut_View.js->onFillTemplateOptions()

+Копирование документа, ИП, ФЛ ошибки отображения

+При продлении контракта начинает считать со дня окончания, а надо + 1 день,
Ипрсавлена функция ExpertiseProlongation->calc_date(), прибавлен день к дате конца работ по контракту

+Переименовал Сроки в списке услу. ServiceList_View.js

+Пустые уведомления. пустая тема в письме
	application_processes_process()
	doc_flow_out_client_process()

+Копирование внеконтрактов попадает в контракт

+Нумерация контракта при переводе во внеконтракт
-Непонятные заблокированные разделы в ответах

+++ functions/file_upload.php
@@ -51,11 +51,11 @@ function mkdir_or_error($dir){
 	if (!file_exists($dir)){
 		if (strlen($dir)>DIR_MAX_LENGTH){
 			throw_common_error('file_uploader Path lenght exceeds maximum value!');
-		}
+		}		
 		if(@mkdir($dir,0775,TRUE)!==TRUE){
-			
+			sleep(1);
 			if (file_exists($dir) && is_dir($dir)) {
-				// The directory was created by a concurrent process, so do nothing, keep calm and carry on
+				// The directory was created by a concurrent process, so do nothing
 			} else {
 				$error = error_get_last();
 				throw_common_error('file_uploader mkdir_or_error '.$dir.' error:'.$error['message']);
@@ -277,7 +277,7 @@ function get_doc_flow_out_client_id_for_db($dbLink,$appIdForDb,$docFlowOutClient
 		FROM doc_flow_out_client
 		WHERE id=%d",$appIdForDb,$db_doc_flow_out_client_id
 	));
-	if (!count($docFlowFields) || $docFlowFields['app_checked']!='t'){
+	if (!is_array($docFlowFields) || !count($docFlowFields) || $docFlowFields['app_checked']!='t'){
 		error_log('file_uploader, function get_doc_flow_out_client_id_for_db, checking not passed, application='.$appIdForDb);
 		throw new Exception(ER_NO_DOC);
 	}
@@ -491,7 +491,7 @@ try{
 						$db_file_id
 						));
 			
-						if (count($ar) && $ar['signed']=='t'){
+						if (is_array($ar) && count($ar) && $ar['signed']=='t'){
 							throw new Exception(ER_SIGNED);
 						}
 									
@@ -557,7 +557,8 @@ try{
 								$db_app_id,$db_doc_type,$db_file_path,$db_fileName
 								));
 								if (
-									count($ar) && $ar['present']=='t'
+									is_array($ar)
+									&& count($ar) && $ar['present']=='t'
 									&& $ar['file_id']!=$par_file_id
 									&& (!isset($_REQUEST['original_file_id']) || $ar['file_id']!=$_REQUEST['original_file_id'])
 								){
@@ -594,6 +595,10 @@ try{
 								$db_file_id,$db_doc_flow_out_client_id
 								));
 							}
+							//Если при COMMIT возникнет ошибка при ответах на замечания - все сломается!
+							//исходный файл потеряли, а нового нет
+							//Поэтому - сначала COMMIT, потом физическое удаление!!!
+							$dbLink->query('COMMIT');
 							
 							if (isset($_REQUEST['original_file_id'])&&isset($_REQUEST['doc_flow_out_client_id'])){
 								//Загружен новый файл с подписью - удаление оригинального файлы, который заменили
@@ -610,41 +615,18 @@ try{
 									//throw new Exception('Deleting original file '.$db_original_file_id);
 									DocFlowOutClient_Controller::removeOriginalFile($dbLink,$db_original_file_id,$db_file_id,$db_doc_flow_out_client_id);
 								}
-								/*
-								$db_original_file_id = NULL;
-								FieldSQLString::formatForDb($dbLink,$_REQUEST['original_file_id'],$db_original_file_id);
-								if ($db_original_file_id!='null'){
-									//if uploaded by this same document - actual unlinking!
-									$ar = $dbLink->query_first(sprintf(		
-									"SELECT TRUE AS present
-									FROM doc_flow_out_client_document_files
-									WHERE file_id=%s AND doc_flow_out_client_id=%d",
-									$db_original_file_id,$db_doc_flow_out_client_id
-									));
-									$unlink_file = (count($ar) && $ar['present']=='t');
-									Application_Controller::removeFile($dbLink,$db_original_file_id,$unlink_file);
-							
-									if (!$unlink_file){
-										$dbLink->query(sprintf(		
-										"INSERT INTO doc_flow_out_client_document_files
-										(file_id,doc_flow_out_client_id,is_new)
-										VALUES (%s,%d,FALSE)
-										ON CONFLICT DO NOTHING",
-										$db_original_file_id,$db_doc_flow_out_client_id
-										));
-									}
-								}
-								*/
 							}
-							
-							$dbLink->query('COMMIT');
 						}	
 						catch(Exception $e){
 							$dbLink->query('ROLLBACK');
 							
 							//Косяк БД - удалим файл и подпись
-							unlink($file_doc);
-							unlink($file_doc_sig);						
+							if(file_exists($file_doc)){
+								unlink($file_doc);
+							}
+							if(file_exists($file_doc_sig)){
+								unlink($file_doc_sig);
+							}
 							
 							throw $e;
 						}
@@ -661,7 +643,7 @@ try{
 							Application_Controller::LKPostfix(),
 							$db_file_id
 							));
-							if (!count($ar) || $ar['present']!='t'){
+							if (!is_array($ar) || !count($ar) || $ar['present']!='t'){
 								check_signature($dbLink,$file_doc,$file_doc_sig,$db_file_id);
 							}
 						}
@@ -728,7 +710,7 @@ try{
 			
 			$ar = $dbLink->query_first($ar_q);
 			
-			if (!count($ar)){
+			if (!is_array($ar) || !count($ar)){
 				throw new Exception(DocFlow_Controller::ER_NOT_FOUND);
 			}
 			else if ($ar['to_application_id']){
@@ -808,7 +790,7 @@ try{
 					WHERE file_id=%s",
 				$db_file_id
 				));
-				$file_exists_in_db = (count($fl_ar) && $fl_ar['file_exists']=='t');
+				$file_exists_in_db = (is_array($fl_ar) && count($fl_ar) && $fl_ar['file_exists']=='t');
 				
 				if (!$sig_add && !$is_sig){
 					//data file
@@ -834,9 +816,11 @@ try{
 	
 						$dbLink->query(sprintf(
 						"INSERT INTO doc_flow_attachments
-						(file_id,doc_type,doc_id,file_size,file_name,file_path,file_signed,employee_id)
+						(file_id,doc_type,doc_id,file_size,file_name,file_path,file_signed,employee_id,require_client_sig)
 						VALUES
-						(%s,'%s'::data_types,%d,%s,%s,%s,%s,%d)",
+						(%s,'%s'::data_types,%d,%s,%s,%s,%s,%d,
+						(SELECT fld.require_client_sig FROM application_doc_folders AS fld
+						WHERE fld.name=%s))",
 							$db_file_id,
 							$doc_type,
 							$db_id,			
@@ -844,7 +828,8 @@ try{
 							$db_file_name,
 							$db_file_path,
 							(isset($_REQUEST['file_signed']) && $_REQUEST['file_signed']=='true')? 'TRUE':'FALSE',
-							json_decode($_SESSION['employees_ref'])->keys->id
+							json_decode($_SESSION['employees_ref'])->keys->id,
+							$db_file_path
 						));
 					
 						if ($doc_type=='doc_flow_out'){
